GSYE-392: Rename aggregated_profile.py to aggregated_ssp_profile.py
