+++ gsy_framework/forward_markets/aggregated_ssp_profile.py
